backend/requirements.txt
flask
flask_cors
requests
werkzeug
gunicorn
python-dotenv